LayoutTests/imported/w3c/web-platform-tests/css/css-transitions/transition-base-response-001-expected.txt


FAIL em units respond to font-size transition assert_equals: expected "15px" but got "20px"
FAIL ex units respond to font-size transition assert_equals: expected "6.71875px" but got "8.96875px"
PASS var() references respond to custom property transition
